# crates/crucible-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crucible web API",
    "description": "The HTTP API that `cru web` serves. This document is generated from the axum router; do not edit it by hand.",
    "version": "0.30.0"
  },
  "paths": {
...
    "/api/skills": {
      "get": {
        "summary": "The skills a kiln discovers, optionally narrowed to one scope.",
        "operationId": "list_skills",
        "parameters": [
          {
            "name": "kiln",
            "in": "query",
            "description": "The kiln to discover skills from.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "description": "Keep only the skills this discovery scope found.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillsReply"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not discover the skills"
          }
        }
      }
    },
    "/api/skills/search": {
      "get": {
        "summary": "The same rows `GET /api/skills` answers, narrowed by a query string.",
        "operationId": "search_skills",
        "parameters": [
          {
            "name": "kiln",
            "in": "query",
            "description": "The kiln to search.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The query, matched case-insensitively against name and description.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many rows to answer with. The daemon's own default applies when\nthe caller names none.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillsReply"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not search the skills"
          }
        }
      }
    },
    "/api/skills/{name}": {
      "get": {
        "summary": "One skill, with the Markdown body a summary row omits.",
        "operationId": "get_skill",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The skill's name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kiln",
            "in": "query",
            "description": "The kiln to discover the skill from.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetail"
                }
              }
            }
          },
          "422": {
            "description": "No skill of that name is discoverable from this kiln"
          },
          "502": {
            "description": "The daemon could not read the skill"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SkillDetail": {
        "type": "object",
        "description": "What `skills.get` answers: one skill, with the body a summary omits.",
        "required": [
          "name",
          "scope",
          "description",
          "source_path",
          "agent",
          "license",
          "body"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "description": "The discovery scope the skill came from, as `SkillScope` spells it."
          },
          "description": {
            "type": "string"
          },
          "source_path": {
            "type": "string",
            "description": "Where the skill file sits on disk."
          },
          "agent": {
            "type": [
              "string",
              "null"
            ],
            "description": "The agent the skill declares, when it declares one. Always written."
          },
          "license": {
            "type": [
              "string",
              "null"
            ],
            "description": "The licence the skill declares, when it declares one. Always written."
          },
          "body": {
            "type": "string",
            "description": "The skill's Markdown body, without its frontmatter."
          }
        }
      },
      "SkillSummary": {
        "type": "object",
        "description": "One skill in a `skills.list` or `skills.search` answer.\n\nThe two RPCs answer the same row, because a list and a search are the same\nquestion asked of two different sets. A second row type here would let one\nof them grow a field the other cannot report.",
        "required": [
          "name",
          "scope",
          "description",
          "shadowed_count"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The skill's name, which is also the key `skills.get` takes."
          },
          "scope": {
            "type": "string",
            "description": "The discovery scope the skill came from, as `SkillScope` spells it."
          },
          "description": {
            "type": "string"
          },
          "shadowed_count": {
            "type": "integer",
            "description": "How many same-named skills this one shadows.",
            "minimum": 0
          }
        }
      },
      "SkillsReply": {
        "type": "object",
        "description": "What `skills.list` and `skills.search` answer.",
        "required": [
          "skills"
        ],
        "properties": {
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillSummary"
            }
          }
        }
      },
...
    }
  }
}